Crate java_game

[J]ust [a]nother [v]irtual [a]dventure game

Originating from the Software Praktikum at TU Dortmund and written in Java, this is a Rust rewrite of the computer game based on the board game "The Forbidden Island". It is purely for educational purposes and fun.

Modules

action_state
Action State of the currently active turn. A turn starts with the player's adventurer taking up to three actions. Afterwards, they may draw two cards from the treasure deck cards and finally the appropriate amount of flood cards depending on the current water level.

adventurer
This module contains the implementation of the adventurers embarking on the forbidden island on the base trait of Adventurer and adventurer info. There are six in the standard game. For information on how they work, please see the corresponding adventurer's module.

artefact_type

asset

bfs
Implementation of a special Breadth First Search algorithm specialised on the applications in the forbidden island.

card_stack
Data type used to store the treasure cards and the flood cards respectively.

character

difficulty

direction

iter_2d
Traits for iterating through a 2D-Vector.

map
Mapse contain the playing field of adventure island.

math

positionable
Items that can be placed on a map, either full or black and white should implement the Positionable trait.

state
The module containing all states, the game can be in

system

water_level
Water level determines how many islands are flooded in the flood card drawing phase. The beginning water level is the primary tool of setting the difficulty of the game.