Trait java_game::adventurer::Adventurer

pub trait Adventurer: AdventurerInfo + Send + Sync {
    fn can_act(&self, act_points: u8) -> bool { ... }

    fn moves(&self, map: &MapFull, act_points: u8) -> Vec<Vec2<u8>> { ... }

    fn special_moves(&self, map: &MapFull, act_points: u8) -> Vec<Vec2<u8>> { ... }

    fn on_move(&mut self) { ... }

    fn can_move_other(&self, _act_points: u8) -> bool { ... }

    fn on_move_other(&mut self, _act_points: &mut u8) { ... }

    fn drains(&self, map: &MapFull, act_points: u8) -> Vec<Vec2<u8>> { ... }

    fn on_drain(&mut self, act_points: &mut u8) { ... }

    fn can_trade_with(&self, other: &dyn Positionable, act_points: u8) -> bool { ... }
}

The specific actions an adventurer can take, also taking the current state of the turn into account. The standard implementation is usually just the AdventurerInfo trait's implementation if there are action points left or nothing/false if they are zero.

For the specific implementations, please see the corresponding adventurer's module for further information.

Provided methods

fn can_act(&self, act_points: u8) -> bool

Checks if the adventurer still has something they can do this turn and returns true if so, false if they cannot do anything except end their turn.

fn moves(&self, map: &MapFull, act_points: u8) -> Vec<Vec2<u8>>

fn special_moves(&self, map: &MapFull, act_points: u8) -> Vec<Vec2<u8>>

fn on_move(&mut self)

If something special should happen, when the adventurer moves, this needs to be implemented. Does not get the action points, since none of the standard adventurers have extra movements per action point.

fn can_move_other(&self, _act_points: u8) -> bool

fn on_move_other(&mut self, _act_points: &mut u8)

When the adventurer has moved another this is called. Must be implemented, if something special should happen, like with the Navigator, which has to handle their extra push.

fn drains(&self, map: &MapFull, act_points: u8) -> Vec<Vec2<u8>>

fn on_drain(&mut self, act_points: &mut u8)

If something special happens, the adventurer must implement this function. Default implementation is to subtract one from the action points.

fn can_trade_with(&self, other: &dyn Positionable, act_points: u8) -> bool

Loading content...