Function java_game::bfs::bfs

pub fn bfs<T, O, M, F, G>(
    data: &Map<T>, 
    start_pos: Option<FieldPos>, 
    start_marker: M, 
    marker: F, 
    output: G
) -> Result<Map<O>, Error> where
    Map<T>: MapExt,
    O: Clone,
    M: Clone + PartialEq,
    F: Fn((FieldPos, &Option<M>), (FieldPos, &T)) -> Option<M>,
    G: Fn(FieldPos, &Option<M>) -> O, 

Breadth first search

Generic general purpose implementation of the breadth first search algorithm. Can be used as a basis for other algorithms that use breadth first search.

Arguments

data - The data to analyse

start_pos - The position from which to start bfs, or None for any starting position that is standable

start_marker - The marker that should be placed at the starting pos of bfs

marker - Function that marks all positions which have been reached based on the tile it has been reached from

output - Function to convert the markers to the expected output