Module java_game::map

Mapse contain the playing field of adventure island.

A valid map contains 24 tiles that are connected in such a way, that the standard moveset of up, down, left and right suffices to walk the entire island by foot.

Re-exports

pub use self::black_white::*;

pub use self::full::*;

pub use self::island_tile::*;

Modules

black_white
The black and white map is the raw map data that just saves, where island tiles should be placed (true) and where the sea is (false). They are usually used to generate maps randomly.

full
Mapse, that can be played on contain 24 unique island tiles.

generate
Utilities to generate maps

island_tile

Structs

Map

Constants

TILE_AMOUNT
The number of tiles that are in a single map when it is valid. In later versions, this could vary.

Traits

MapExt

Type Definitions

FieldPos
Alias for the position type an item must hold in order to be placeable on a map.